*: move meta to internal
